Move types to the graphql-module
